+++ _legacy/components/GoButton/index.tsx
