+++ _legacy/components/GoButton/style.ts
