+++ _legacy/components/Nickname/style.ts
