+++ _legacy/components/Toggle/style.ts
@@ -3,7 +3,3 @@ import styled from "styled-components";
 export const ToggleItem = styled.li`
   list-style: none;
 `;
-
-export const ToggleWrapper = styled.label`
-  font-size: 0.8rem;
-`;
