+++ _legacy/components/favoriteChannel/index.tsx
