+++ _legacy/components/favoriteChannel/style.ts
