+++ package.json
@@ -11,23 +11,27 @@
   "license": "ISC",
   "devDependencies": {
     "@swc/core": "^1.3.102",
+    "@types/file-saver": "^2.0.7",
     "@types/react": "^18.2.47",
     "@types/react-dom": "^18.2.18",
-    "css-loader": "^6.9.0",
+    "css-loader": "^7.1.2",
     "file-loader": "^6.2.0",
+    "sass-loader": "^16.0.3",
     "style-loader": "^3.3.4",
     "swc-loader": "^0.2.3",
     "webpack": "^5.89.0",
+    "sass": "1.77.6",
     "webpack-cli": "^5.1.4"
   },
   "dependencies": {
+    "@picocss/pico": "^2.0.6",
     "@types/chrome": "^0.0.256",
     "axios": "^1.6.5",
+    "bootstrap": "^5.3.3",
+    "file-saver": "^2.0.5",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.10.0",
+    "react-bootstrap": "^2.10.5",
     "react-dom": "^18.2.0",
-    "sass": "^1.69.7",
-    "sass-loader": "^13.3.3",
     "styled-components": "^6.1.8"
   },
   "repository": {
