+++ public/index.html
@@ -3,15 +3,13 @@
 <head>
     <meta charset="UTF-8"/>
     <link
-            rel="stylesheet"
-            href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-            integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-            crossorigin="anonymous"
-    />
+        rel="stylesheet"
+        href="https://cdn.jsdelivr.net/npm/@picocss/pico@2/css/pico.min.css"
+    >
     <title><%=mode%></title>
 </head>
 <body style="background: #0c0d0e">
-<main id="root"></main>
+<main id="root" style="padding: 0;"></main>
 <script src="./main.js"></script>
 </body>
 </html>
