+++ src/App.css
@@ -1,0 +1,85 @@
+.wrapper {
+  text-align: center;
+  width: 350px;
+  display: flex;
+  justify-content: start;
+  flex-direction: column;
+  border: 1px solid #28292a;
+  margin: 0;
+}
+
+header {
+  display: flex;
+  align-items: flex-end;  
+  justify-content: space-between;
+  padding-top: 1.5rem!important;
+
+  img {
+    @media (prefers-color-scheme: dark) {
+      content: url('/logo-dark.png');
+    }
+
+    @media (prefers-color-scheme: light) {
+      content: url('/logo-light.png');
+    }
+  }
+
+  .top-menu {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    text-align: right;
+    a{
+      font-size: 0.7rem;
+      text-decoration: underline;
+    }
+  }
+}
+
+footer {
+  background: #f5f5f5;
+  height: 5rem;
+  /* border-top: 1px solid #28292a; */
+}
+
+.filter-list {
+  fieldset[role="group"] {
+    input {
+      --pico-form-element-spacing-vertical: 0rem;
+      --pico-form-element-spacing-horizontal: 0.2rem;
+    }
+    button {
+      --pico-form-element-spacing-vertical: 0rem;
+      --pico-form-element-spacing-horizontal: 0.2rem;
+      font-size: 0.7rem;
+      flex: 0 0 100px;
+    }
+    margin-bottom: 0.25rem;
+  }
+
+  details[role="button"] {
+    cursor: default;
+
+    summary {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      text-align: left;
+      line-height: 1.5;
+    }
+    .filter-item {
+      cursor: pointer;
+      padding: 0.25rem;
+      background: #f5f5f5;
+      margin-bottom: 0.25rem;
+      border-radius: var(--pico-border-radius);
+      background-color: var(--pico-primary-background);
+      color: var(--pico-primary-inverse);
+      padding: 0.1rem 0.75rem;
+      text-align: left;
+    }
+    .filter-item:hover {
+      background: var(--pico-primary-hover-background);
+    }
+  }
+}
