+++ src/components/SubTitle/index.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const Divider = styled.div`
+  font-size: 1rem;
+  text-align: left;
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+`;
+
+export default Divider;
