+++ src/components/Toggle/index.tsx
@@ -1,19 +1,41 @@
-import { ToggleProps } from "@/src/types/toggle";
-import React from "react";
-import { Form } from "react-bootstrap";
-import { ToggleItem } from "./style";
+
+import { ToggleProps } from "../../types/toggle";
+import { ToggleWrapper } from "./style";
 
 const Toggle = ({ onChange, label, value }: ToggleProps) => {
   return (
-    <ToggleItem>
-      <Form.Check
+    <ToggleWrapper>
+      <input
         onChange={onChange}
-        type="switch"
-        label={label}
+        type="checkbox" role="switch" 
         checked={value}
       />
-    </ToggleItem>
+      {label}
+    </ToggleWrapper>
   );
 };
 
 export default Toggle;
+
+
+
+
+// import { ToggleProps } from "@/src/types/toggle";
+// import React from "react";
+// import { Form } from "react-bootstrap";
+// import { ToggleItem } from "./style";
+
+// const Toggle = ({ onChange, label, value }: ToggleProps) => {
+//   return (
+//     <ToggleItem>
+//       <Form.Check
+//         onChange={onChange}
+//         type="switch"
+//         label={label}
+//         checked={value}
+//       />
+//     </ToggleItem>
+//   );
+// };
+
+// export default Toggle;
